fix: merge common methods
