+++ package.json
@@ -23,7 +23,7 @@
   },
   "dependencies": {
     "@auth/prisma-adapter": "^2.7.4",
-    "@aws-sdk/client-s3": "^3.750.0",
+    "@aws-sdk/client-s3": "3.830.0",
     "@aws-sdk/s3-request-presigner": "^3.810.0",
     "@hono/node-server": "^1.13.8",
     "@hookform/resolvers": "^3.10.0",
