+++ server/lib/s3.ts
@@ -1,5 +1,4 @@
-import { S3Client, PutObjectCommand, GetObjectCommand } from '@aws-sdk/client-s3'
-import { getSignedUrl } from '@aws-sdk/s3-request-presigner'
+import { S3Client } from '@aws-sdk/client-s3'
 import type { Config } from '~/types'
 
 let s3Client: S3Client | null = null
@@ -38,24 +37,4 @@ export function getClient(findConfig: Config[]) {
   }
 
   return s3Client
-}
-
-export async function generatePresignedUrl(
-  s3Client: S3Client,
-  bucket: string,
-  key: string,
-  operation: 'get' | 'put' = 'get',
-  expiresIn: number = 3600,
-) {
-  const command = operation === 'get'
-    ? new GetObjectCommand({
-        Bucket: bucket,
-        Key: key,
-      })
-    : new PutObjectCommand({
-        Bucket: bucket,
-        Key: key,
-      })
-
-  return await getSignedUrl(s3Client, command, { expiresIn })
 }
