+++ server/lib/s3api.ts
@@ -1,20 +1,26 @@
 import 'server-only'
 
-import { PutObjectCommand, S3Client } from '@aws-sdk/client-s3'
+import { GetObjectCommand, PutObjectCommand, S3Client } from '@aws-sdk/client-s3'
+import { getSignedUrl } from '@aws-sdk/s3-request-presigner'
 
-/**
- * 上传单个对象到存储桶
- * @param client s3 客户端
- * @param params 上传参数
- */
-export async function uploadSimpleObject(client: S3Client, params: {
-  Bucket: string,
-  Key: string,
-  Body: Buffer,
-  ContentLength: number
-  ContentType: string
-}) {
-  await client.send(
-    new PutObjectCommand(params)
-  )
+export async function generatePresignedUrl(
+  s3Client: S3Client,
+  bucket: string,
+  key: string,
+  contentType: string,
+  operation: 'get' | 'put' = 'get',
+  expiresIn: number = 3600,
+) {
+  const command = operation === 'get'
+    ? new GetObjectCommand({
+      Bucket: bucket,
+      Key: key,
+    })
+    : new PutObjectCommand({
+      Bucket: bucket,
+      Key: key,
+      ContentType: contentType || undefined
+    })
+
+  return await getSignedUrl(s3Client as any, command as any, { expiresIn })
 }
